Write an evaluation of this approach
Have a look at how I understand how the microservice works and the comparison
at the bottom of the file. Please correct me if I get it wrong and feel free to
answer my open questions. I think the big advantage of this microservice
is that we don't have to settle for a fixed size of image resolutions
but have some sort of microservice which is doing the conversion on the
fly. However I believe this microservice adds a lot of complexity to our
code base. Currently I don't see the justification for it if we can more or
less achieve the same behaviour *without any self-written service*.